feat: implement core protocol functions for borrowing, depositing, repaying, withdrawing, and liquidating assets
