+++ stellar-lend/contracts/hello-world/src/lib.rs
@@ -12,12 +12,12 @@ use soroban_sdk::{
     String, Symbol, Vec,
 };
 
-// Module placeholders for future expansion
-// mod deposit;
-// mod borrow;
-// mod repay;
-// mod withdraw;
-// mod liquidate;
+// Core protocol modules
+mod deposit;
+mod borrow;
+mod repay;
+mod withdraw;
+mod liquidate;
 
 
 /// Reentrancy guard for security
@@ -2286,379 +2286,22 @@ impl Contract {
     // --- Core Protocol Function Placeholders ---
 /// Deposit collateral into the protocol
 pub fn deposit_collateral(env: Env, depositor: String, amount: i128) -> Result<(), ProtocolError> {
-    ReentrancyGuard::enter(&env)?;
-    let result = (|| {
-        // Input validation with enhanced error logging
-        if depositor.is_empty() {
-            let error = ProtocolError::InvalidAddress;
-            ErrorLogger::log_error(&env, &error, None, "deposit_collateral", "Empty depositor address provided");
-            return Err(error);
-        }
-        if amount <= 0 {
-            let error = ProtocolError::InvalidAmount;
-            let context = format!("Invalid deposit amount: {}", amount);
-            ErrorLogger::log_error(&env, &error, None, "deposit_collateral", &context);
-            return Err(error);
-        }
-
-        // Check if deposit is paused
-        let risk_config = RiskConfigStorage::get(&env);
-        if risk_config.pause_deposit {
-            let error = ProtocolError::ProtocolPaused;
-            ErrorLogger::log_error(&env, &error, None, "deposit_collateral", "Deposit operations are paused");
-            return Err(error);
-        }
-
-        let depositor_addr = Address::from_string(&depositor);
-        
-        // Enhanced security checks with error logging
-        if FrozenAccounts::is_frozen(&env, &depositor_addr) {
-            SecurityMonitor::record_suspicious(&env, &depositor_addr, "deposit while frozen");
-            let error = ProtocolError::Unauthorized;
-            ErrorLogger::log_error(&env, &error, Some(depositor_addr.clone()), "deposit_collateral", "Account is frozen");
-            return Err(error);
-        }
-        
-        // Compliance checks with error logging
-        if let Err(error) = require_kyc(&env, &depositor_addr) {
-            ErrorLogger::log_error(&env, &error, Some(depositor_addr.clone()), "deposit_collateral", "KYC requirement not met");
-            return Err(error);
-        }
-        if let Err(error) = require_not_blacklisted(&env, &depositor_addr) {
-            ErrorLogger::log_error(&env, &error, Some(depositor_addr.clone()), "deposit_collateral", "Address is blacklisted");
-            return Err(error);
-        }
-        if let Err(error) = check_aml(&env, &depositor_addr, amount, "deposit") {
-            ErrorLogger::log_error(&env, &error, Some(depositor_addr.clone()), "deposit_collateral", "AML check failed");
-            return Err(error);
-        }
-        // Load user position with error handling
-        let mut position = match StateHelper::get_position(&env, &depositor_addr) {
-            Some(pos) => pos,
-            None => Position::new(depositor_addr.clone(), 0, 0),
-        };
-
-        // Accrue interest before updating position with error handling
-        let state = match InterestRateStorage::update_state(&env) {
-            state => state,
-        };
-        
-        InterestRateManager::accrue_interest_for_position(
-            &env,
-            &mut position,
-            state.current_borrow_rate,
-            state.current_supply_rate,
-        );
-
-        // Update position with error recovery
-        position.collateral += amount;
-        
-        // Attempt to save position with error recovery
-        let save_result = || -> Result<(), ProtocolError> {
-            StateHelper::save_position(&env, &position);
-            Ok(())
-        };
-
-        if let Err(error) = save_result() {
-            let storage_error = ProtocolError::StorageError;
-            let context = ErrorLogger::log_error(
-                &env, 
-                &storage_error, 
-                Some(depositor_addr.clone()), 
-                "deposit_collateral", 
-                "Failed to save user position"
-            );
-            
-            // Attempt recovery
-            match ErrorRecovery::attempt_recovery(&env, &storage_error, context) {
-                Ok(()) => {
-                    // Retry the save operation
-                    StateHelper::save_position(&env, &position);
-                }
-                Err(_) => return Err(storage_error),
-            }
-        }
-
-        // Update total supplied amount with error recovery
-        let mut ir_state = InterestRateStorage::get_state(&env);
-        ir_state.total_supplied += amount;
-        
-        let save_ir_result = || -> Result<(), ProtocolError> {
-            InterestRateStorage::save_state(&env, &ir_state);
-            Ok(())
-        };
-
-        if let Err(_) = save_ir_result() {
-            let storage_error = ProtocolError::StorageError;
-            let context = ErrorLogger::log_error(
-                &env, 
-                &storage_error, 
-                Some(depositor_addr.clone()), 
-                "deposit_collateral", 
-                "Failed to update interest rate state"
-            );
-            
-            // Attempt recovery
-            if let Err(_) = ErrorRecovery::attempt_recovery(&env, &storage_error, context) {
-                return Err(storage_error);
-            }
-        }
-
-        // Collect any accrued supply interest as protocol fees
-        if position.supply_interest > 0 {
-            let config = InterestRateStorage::get_config(&env);
-            let (_, supply_fee) = InterestRateManager::collect_fees_from_interest(
-                &env,
-                0,
-                position.supply_interest,
-                config.reserve_factor,
-            );
-            if supply_fee > 0 {
-                let mut reserve_data = ReserveStorage::get_reserve_data(&env);
-                reserve_data.total_fees_collected += supply_fee;
-                reserve_data.current_reserves += supply_fee;
-                ReserveStorage::save_reserve_data(&env, &reserve_data);
-
-                // Update revenue metrics
-                let mut metrics = ReserveStorage::get_revenue_metrics(&env);
-                metrics.total_supply_fees += supply_fee;
-                ReserveStorage::save_revenue_metrics(&env, &metrics);
-
-                ProtocolEvent::FeesCollected {
-                    amount: supply_fee,
-                    source: String::from_str(&env, "supply"),
-                }
-                .emit(&env);
-            }
-        }
-
-        ProtocolEvent::Deposit {
-            user: depositor,
-            amount,
-            asset: String::from_str(&env, "XLM"),
-        }
-        .emit(&env);
-
-        Ok(())
-    })();
-    ReentrancyGuard::exit(&env);
-    result
+    deposit::deposit_collateral(&env, &depositor.to_string(), amount)
 }
 /// Borrow assets from the protocol with dynamic risk check
 pub fn borrow(env: Env, borrower: String, amount: i128) -> Result<(), ProtocolError> {
-    ReentrancyGuard::enter(&env)?;
-    let result = (|| {
-        if borrower.is_empty() {
-            return Err(ProtocolError::InvalidAddress);
-        }
-        if amount <= 0 {
-            return Err(ProtocolError::InvalidAmount);
-        }
-
-        // Check if borrow is paused
-        let risk_config = RiskConfigStorage::get(&env);
-        if risk_config.pause_borrow {
-            return Err(ProtocolError::ProtocolPaused);
-        }
-
-        let borrower_addr = Address::from_string(&borrower);
-        if FrozenAccounts::is_frozen(&env, &borrower_addr) {
-            SecurityMonitor::record_suspicious(&env, &borrower_addr, "borrow while frozen");
-            return Err(ProtocolError::Unauthorized);
-        }
-        require_kyc(&env, &borrower_addr)?;
-        require_not_blacklisted(&env, &borrower_addr)?;
-        check_aml(&env, &borrower_addr, amount, "borrow")?;
-        let mut position = StateHelper::get_position(&env, &borrower_addr)
-            .unwrap_or(Position::new(borrower_addr.clone(), 0, 0));
-
-        // Accrue interest before updating position
-        let state = InterestRateStorage::update_state(&env);
-        InterestRateManager::accrue_interest_for_position(
-            &env,
-            &mut position,
-            state.current_borrow_rate,
-            state.current_supply_rate,
-        );
-
-        let new_debt = position.debt + amount;
-        let mut new_position = position.clone();
-        new_position.debt = new_debt;
-
-        let min_ratio = ProtocolConfig::get_min_collateral_ratio(&env);
-        let ratio = StateHelper::dynamic_collateral_ratio::<RealPriceOracle>(&env, &new_position);
-        if ratio < min_ratio {
-            SecurityMonitor::record_suspicious(&env, &borrower_addr, "borrow below collateral ratio");
-            return Err(ProtocolError::InsufficientCollateralRatio);
-        }
-
-        position.debt = new_debt;
-        StateHelper::save_position(&env, &position);
-
-        // Update total borrowed amount
-        let mut ir_state = InterestRateStorage::get_state(&env);
-        ir_state.total_borrowed += amount;
-        InterestRateStorage::save_state(&env, &ir_state);
-
-        // Collect any accrued borrow interest as protocol fees
-        if position.borrow_interest > 0 {
-            let config = InterestRateStorage::get_config(&env);
-            let (borrow_fee, _) = InterestRateManager::collect_fees_from_interest(
-                &env,
-                position.borrow_interest,
-                0,
-                config.reserve_factor,
-            );
-            if borrow_fee > 0 {
-                let mut reserve_data = ReserveStorage::get_reserve_data(&env);
-                reserve_data.total_fees_collected += borrow_fee;
-                reserve_data.current_reserves += borrow_fee;
-                ReserveStorage::save_reserve_data(&env, &reserve_data);
-
-                // Update revenue metrics
-                let mut metrics = ReserveStorage::get_revenue_metrics(&env);
-                metrics.total_borrow_fees += borrow_fee;
-                ReserveStorage::save_revenue_metrics(&env, &metrics);
-
-                ProtocolEvent::FeesCollected {
-                    amount: borrow_fee,
-                    source: String::from_str(&env, "borrow"),
-                }
-                .emit(&env);
-            }
-        }
-
-        ProtocolEvent::Borrow {
-            user: borrower,
-            amount,
-            asset: String::from_str(&env, "XLM"),
-        }
-        .emit(&env);
-
-        Ok(())
-    })();
-    ReentrancyGuard::exit(&env);
-    result
+    borrow::borrow(&env, &borrower.to_string(), amount)
 }
 
 /// Repay borrowed assets
 pub fn repay(env: Env, repayer: String, amount: i128) -> Result<(), ProtocolError> {
-    ReentrancyGuard::enter(&env)?;
-    let result = (|| {
-        if repayer.is_empty() {
-            return Err(ProtocolError::InvalidAddress);
-        }
-        if amount <= 0 {
-            return Err(ProtocolError::InvalidAmount);
-        }
-
-        let repayer_addr = Address::from_string(&repayer);
-        if FrozenAccounts::is_frozen(&env, &repayer_addr) {
-            SecurityMonitor::record_suspicious(&env, &repayer_addr, "repay while frozen");
-            return Err(ProtocolError::Unauthorized);
-        }
-        require_kyc(&env, &repayer_addr)?;
-        require_not_blacklisted(&env, &repayer_addr)?;
-        check_aml(&env, &repayer_addr, amount, "repay")?;
-        let mut position = StateHelper::get_position(&env, &repayer_addr).unwrap_or(Position::new(
-            repayer_addr.clone(),
-            0,
-            0,
-        ));
-
-        let mut position = StateHelper::get_position(&env, &repayer_addr)
-            .unwrap_or(Position::new(repayer_addr.clone(), 0, 0));
-
-        // Accrue interest before updating position
-        let state = InterestRateStorage::update_state(&env);
-        InterestRateManager::accrue_interest_for_position(
-            &env,
-            &mut position,
-            state.current_borrow_rate,
-            state.current_supply_rate,
-        );
-
-        let old_debt = position.debt;
-        position.debt = (position.debt - amount).max(0);
-        StateHelper::save_position(&env, &position);
-
-        // Update total borrowed amount
-        let mut ir_state = InterestRateStorage::get_state(&env);
-        ir_state.total_borrowed -= (old_debt - position.debt);
-        InterestRateStorage::save_state(&env, &ir_state);
-
-        ProtocolEvent::Repay {
-            user: repayer,
-            amount,
-            asset: String::from_str(&env, "XLM"),
-        }
-        .emit(&env);
-
-        Ok(())
-    })();
-    ReentrancyGuard::exit(&env);
-    result
+    repay::repay(&env, &repayer.to_string(), amount)
 }
 
 
     /// Withdraw collateral with dynamic risk check
     pub fn withdraw(env: Env, withdrawer: String, amount: i128) -> Result<(), ProtocolError> {
-        if withdrawer.is_empty() {
-            return Err(ProtocolError::InvalidAddress);
-        }
-        if amount <= 0 {
-            return Err(ProtocolError::InvalidAmount);
-        }
-
-        // Check if withdraw is paused
-        let risk_config = RiskConfigStorage::get(&env);
-        if risk_config.pause_withdraw {
-            return Err(ProtocolError::ProtocolPaused);
-        }
-        let withdrawer_addr = Address::from_string(&withdrawer);
-        if FrozenAccounts::is_frozen(&env, &withdrawer_addr) {
-            return Err(ProtocolError::Unauthorized);
-        }
-        require_kyc(&env, &withdrawer_addr)?;
-        require_not_blacklisted(&env, &withdrawer_addr)?;
-        check_aml(&env, &withdrawer_addr, amount, "withdraw")?;
-        let mut position = StateHelper::get_position(&env, &withdrawer_addr)
-            .unwrap_or(Position::new(withdrawer_addr.clone(), 0, 0));
-
-        // Accrue interest before updating position
-        let state = InterestRateStorage::update_state(&env);
-        InterestRateManager::accrue_interest_for_position(
-            &env,
-            &mut position,
-            state.current_borrow_rate,
-            state.current_supply_rate,
-        );
-
-        if position.collateral < amount {
-            return Err(ProtocolError::InsufficientCollateral);
-        }
-        let mut new_position = position.clone();
-        new_position.collateral -= amount;
-        let min_ratio = ProtocolConfig::get_min_collateral_ratio(&env);
-        let ratio = StateHelper::dynamic_collateral_ratio::<RealPriceOracle>(&env, &new_position);
-        if position.debt > 0 && ratio < min_ratio {
-            return Err(ProtocolError::InsufficientCollateralRatio);
-        }
-        position.collateral = new_position.collateral;
-        StateHelper::save_position(&env, &position);
-
-        // Update total supplied amount
-        let mut ir_state = InterestRateStorage::get_state(&env);
-        ir_state.total_supplied -= amount;
-        InterestRateStorage::save_state(&env, &ir_state);
-
-        ProtocolEvent::Withdraw {
-            user: withdrawer,
-            amount,
-            asset: String::from_str(&env, "XLM"),
-        }
-        .emit(&env);
-        Ok(())
+        withdraw::withdraw(&env, &withdrawer.to_string(), amount)
     }
 
     /// Liquidate undercollateralized positions using dynamic risk check
@@ -2668,76 +2311,7 @@ pub fn repay(env: Env, repayer: String, amount: i128) -> Result<(), ProtocolErro
         target: String,
         amount: i128,
     ) -> Result<(), ProtocolError> {
-        if liquidator.is_empty() || target.is_empty() {
-            return Err(ProtocolError::InvalidAddress);
-        }
-        if amount <= 0 {
-            return Err(ProtocolError::InvalidAmount);
-        }
-
-        // Check if liquidation is paused
-        let risk_config = RiskConfigStorage::get(&env);
-        if risk_config.pause_liquidate {
-            return Err(ProtocolError::ProtocolPaused);
-        }
-
-        let target_addr = Address::from_string(&target);
-        if FrozenAccounts::is_frozen(&env, &target_addr) {
-            return Err(ProtocolError::Unauthorized);
-        }
-
-        let mut position = match StateHelper::get_position(&env, &target_addr) {
-            Some(pos) => pos,
-            None => return Err(ProtocolError::PositionNotFound),
-        };
-
-        // Accrue interest before liquidation
-        let state = InterestRateStorage::update_state(&env);
-        InterestRateManager::accrue_interest_for_position(
-            &env,
-            &mut position,
-            state.current_borrow_rate,
-            state.current_supply_rate,
-        );
-
-        let min_ratio = ProtocolConfig::get_min_collateral_ratio(&env);
-        let ratio = StateHelper::dynamic_collateral_ratio::<RealPriceOracle>(&env, &position);
-        if ratio >= min_ratio {
-            return Err(ProtocolError::NotEligibleForLiquidation);
-        }
-
-        // Apply close factor to limit liquidation amount
-        let max_repay_amount = (position.debt * risk_config.close_factor) / 100_000_000;
-        let repay_amount = amount.min(position.debt).min(max_repay_amount);
-
-        if repay_amount == 0 {
-            return Err(ProtocolError::InvalidAmount);
-        }
-
-        // Calculate liquidation incentive
-        let incentive_amount = (repay_amount * risk_config.liquidation_incentive) / 100_000_000;
-        let total_collateral_seized = repay_amount + incentive_amount;
-
-        // Ensure we don't seize more collateral than available
-        let actual_collateral_seized = total_collateral_seized.min(position.collateral);
-
-        // Update position
-        position.debt -= repay_amount;
-        position.collateral -= actual_collateral_seized;
-        StateHelper::save_position(&env, &position);
-
-        // Update total borrowed amount
-        let mut ir_state = InterestRateStorage::get_state(&env);
-        ir_state.total_borrowed -= repay_amount;
-        InterestRateStorage::save_state(&env, &ir_state);
-
-        ProtocolEvent::Liquidate {
-            user: target,
-            amount: repay_amount,
-            asset: String::from_str(&env, "XLM"),
-        }
-        .emit(&env);
-        Ok(())
+        liquidate::liquidate(&env, &liquidator.to_string(), &target.to_string(), amount)
     }
 
     pub fn hello(env: Env, to: String) -> Vec<String> {
@@ -2848,60 +2422,7 @@ pub fn repay(env: Env, repayer: String, amount: i128) -> Result<(), ProtocolErro
         ProtocolConfig::require_admin(&env, &caller_addr)?;
         let mut config = RiskConfigStorage::get(&env);
 
-    /// Withdraw collateral from the protocol
-    pub fn withdraw(env: Env, withdrawer: String, amount: i128) -> Result<(), ProtocolError> {
-        ReentrancyGuard::enter(&env)?;
-        let result = (|| {
-            if withdrawer.is_empty() {
-                return Err(ProtocolError::InvalidAddress);
-            }
-            if amount <= 0 {
-                return Err(ProtocolError::InvalidAmount);
-            }
-            // Check if withdraw is paused
-            let risk_config = RiskConfigStorage::get(&env);
-            if risk_config.pause_withdraw {
-                return Err(ProtocolError::ProtocolPaused);
-            }
-            let withdrawer_addr = Address::from_string(&withdrawer);
-            if FrozenAccounts::is_frozen(&env, &withdrawer_addr) {
-                return Err(ProtocolError::Unauthorized);
-            }
-            // Load user position
-            let mut position = StateHelper::get_position(&env, &withdrawer_addr)
-                .ok_or(ProtocolError::PositionNotFound)?;
-            // Check sufficient collateral
-            if position.collateral < amount {
-                SecurityMonitor::record_suspicious(&env, &withdrawer_addr, "Insufficient collateral for withdrawal");
-                return Err(ProtocolError::InsufficientCollateral);
-            }
-            // Simulate withdrawal and check collateral ratio
-            position.collateral -= amount;
-            let min_ratio = ProtocolConfig::get_min_collateral_ratio(&env);
-            let ratio = StateHelper::collateral_ratio(&position);
-            if ratio < min_ratio {
-                SecurityMonitor::record_suspicious(&env, &withdrawer_addr, "Collateral ratio too low after withdrawal");
-                return Err(ProtocolError::InsufficientCollateralRatio);
-            }
-            // Save updated position
-            StateHelper::save_position(&env, &position);
-            // Track user activity
-            let mut activity = ActivityStorage::get_user_activity(&env, &withdrawer_addr)
-                .unwrap_or(UserActivity::new());
-            activity.record_withdrawal(amount, env.ledger().timestamp());
-            ActivityStorage::save_user_activity(&env, &withdrawer_addr, &activity);
-            // Emit event
-            ProtocolEvent::Withdraw {
-                user: withdrawer.clone(),
-                amount,
-                asset: String::from_str(&env, "XLM"),
-            }
-            .emit(&env);
-            Ok(())
-        })();
-        ReentrancyGuard::exit(&env);
-        result
-    }
+
         config.pause_borrow = pause_borrow;
         config.pause_deposit = pause_deposit;
         config.pause_withdraw = pause_withdraw;
